Complete Spring Boot 2.7.18 migration for opencbs-core with systematic API compatibility fixes (still need fixing on unit test and another module)

+++ server/opencbs-core/src/main/java/com/opencbs/core/audit/reports/AuditEventsReport.java
@@ -48,8 +48,8 @@ public class AuditEventsReport implements AuditReport {
             return new PageImpl(Collections.EMPTY_LIST, pageable, eventIds.size());
         }
 
-        int fromPosition = (eventIds.size() < pageable.getOffset()) ? eventIds.size() : pageable.getOffset();
-        int toPosition = (eventIds.size() > pageable.getOffset() + pageable.getPageSize()) ? pageable.getOffset() + pageable.getPageSize() : eventIds.size();
+        int fromPosition = (eventIds.size() < (int)pageable.getOffset()) ? eventIds.size() : (int)pageable.getOffset();
+        int toPosition = (eventIds.size() > (int)pageable.getOffset() + pageable.getPageSize()) ? (int)pageable.getOffset() + pageable.getPageSize() : eventIds.size();
         List<AuditEventIdentificator> auditEventIdentificators = eventIds.subList(fromPosition, toPosition);
 
         return new PageImpl(getRecordsByIds(auditEventIdentificators), pageable, eventIds.size());
